% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{bhAdjust}
\alias{bhAdjust}
\title{p-value thresholding with Bonferroni-Holm correction}
\usage{
bhAdjust(p)
}
\arguments{
\item{p}{A numeric vector of p-values from an analysis.}
}
\value{
A logical vector indicating whether or not each of the p-values satisfy an alpha of 0.05, following Bonferroni-Holm correction (0 is no, 1 is yes).
}
\description{
Adjusts a set of p-values using the Bonferroni-Holm correction, and returns a logical vector indicating whether each corresponding adjusted p-value is below 0.05 (alpha of 0.05).
}
\examples{
p=seq(0.0025,0.0250,0.0025)
names(p)=paste('Test',LETTERS[seq(1,length(p))])
bhAdjust(p)

}
