% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{r2}
\alias{r2}
\title{R-squared calculation}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{A numeric vector of variable values predicted by the model.}

\item{truth}{A numeric vector of variable values measured.}
}
\value{
The R-squared value for that model.
}
\description{
Calculates the R-squared for a given model, i.e. the measure of the model fit.
}
\examples{
data=read.csv('https://jlucasmckay.bmi.emory.edu/global/bmi585/credit.csv')
lim_inccards=lm(Limit~Income+Cards,data)
R_squared_calc(data$Limit,lim_inccards$fitted.values)

}
