% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{twoSidedZ}
\alias{twoSidedZ}
\title{Conversion of test statistic to z-density}
\usage{
twoSidedZ(z)
}
\arguments{
\item{z}{The test statistic.}
}
\value{
The area under the z-density for values >=|x|.
}
\description{
Converts a test statistic z to the area under the z-density for values >=|x|.
}
\examples{
twoSidedZ(3)

}
